fix(frontend): address remaining react-doctor issues and PR review comments
- Extract provider fetch into useCatalogDefaults hook (fixes last
  fetch-in-useEffect error)
- Move inner components to separate files so react-doctor detects
  Suspense boundaries for useSearchParams
- Split app/page.tsx into server page with metadata + client component
- Fix Conversation.id and ApiConversation.id types from number to string
  to match backend UUID schema
- Change nested nav to div in AppSidebar to avoid duplicate landmarks
- Replace role="button" on UploadMenu backdrop with aria-hidden
- Wrap setMessages in useCallback for stable identity
