+++ frontend/plugins/chat/components/input/UploadMenu.tsx
@@ -19,16 +19,7 @@ export function UploadMenu({
   const inputRef = useRef<HTMLInputElement>(null);
   return (
     <>
-      <div
-        role="button"
-        tabIndex={0}
-        aria-label="Close upload menu"
-        className="fixed inset-0 z-0"
-        onClick={onClose}
-        onKeyDown={(e) => {
-          if (e.key === "Escape") onClose();
-        }}
-      />
+      <div aria-hidden="true" className="fixed inset-0 z-0" onClick={onClose} />
 
       <input
         ref={inputRef}
